requirements.txt
fastapi
python-dotenv
requests
line-bot-sdk
uvicorn[standard]
openai
